// src/app/core/interfaces/product.interface.ts
// Generated by https://quicktype.io

export interface Product {
    total:    number;
    products: ProductElement[];
}

export interface ProductElement {
    _id:         string;
    sku:         string;
    name:        string;
    description: string;
    price:       number;
    stock:       number;
    tag?:       string;
    state:       boolean;
    image?:      string;
    updatedAt: string;
}